Remove extraneous CompositeCheckoutContainer
